Fixed some compiler warnings

+++ src/ColumnString.cpp
@@ -234,8 +234,8 @@ size_t AdaptiveStringColumn::Find(const char* value) const {
 size_t AdaptiveStringColumn::Find(const char* value, size_t len) const {
 	assert(value);
 
-	if (m_len == 0) return -1; // empty list
-	if (len >= m_width) return -1; // A string can never be wider than the column width
+	if (m_len == 0) return (size_t)-1; // empty list
+	if (len >= m_width) return (size_t)-1; // A string can never be wider than the column width
 
 	if (m_width == 0) {
 		return 0; 
@@ -291,7 +291,7 @@ size_t AdaptiveStringColumn::Find(const char* value, size_t len) const {
 	}
 	else assert(false);
 		
-	return -1;
+	return (size_t)-1;
 }
 
 void AdaptiveStringColumn::Stats() const {
